group open poll and invalidate into hook
